requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-PyMongo==2.3.0
flask-restx==1.3.0
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
packaging==23.2
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.1
rpds-py==0.16.2
typing_extensions==4.9.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
